fmt fix

+++ tests/test_circular_buffer.cpp
@@ -22,6 +22,7 @@ TEST_CASE("circular buffer test","")
       {
         REQUIRE(i == 4);
       }
+
       auto writer = b1.begin();
       auto reader = writer.new_reader();
       for(auto i:reader)
@@ -45,19 +46,24 @@ TEST_CASE("circular buffer test","")
   SECTION("basice circular buffer test size 8")
   {
       circular_buffer<uint8_t,8> b1;
+
       b1.emplace_back(1);
       b1.emplace_back(2);
+
       int idx = 1;
+
       for(auto i:b1)
       {
         REQUIRE(i == idx);
         ++idx;
       }
+
       REQUIRE(idx == 3);
       auto loc = b1.begin();
-      REQUIRE((int)*loc == 1);
+      REQUIRE(*loc == 1);
       ++loc;
-      REQUIRE((int)*loc == 2);
+      REQUIRE(*loc == 2);
+      
   }
   SECTION("basice circular buffer test size 8")
   {
